+++ packages/components-js/tests/e2e/select-wrapper.e2e.ts
@@ -1,17 +1,34 @@
-import { newE2EPage } from '@stencil/core/testing';
+import { E2EPage, newE2EPage } from '@stencil/core/testing';
 
 describe('select-wrapper', () => {
   it('should render', async () => {
     const page = await newE2EPage();
-    await page.setContent(`<p-select-wrapper label="Some label">
-      <select name="some-name">
-        <option value="a">Option A</option>
-        <option value="b">Option B</option>
-        <option value="c">Option C</option>
-      </select>
-    </p-select-wrapper>`);
+    await page.setContent(`
+      <p-select-wrapper label="Some label">
+        <select name="some-name">
+          <option value="a">Option A</option>
+          <option value="b">Option B</option>
+          <option value="c">Option C</option>
+        </select>
+      </p-select-wrapper>
+    `);
     const el = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
     expect(el).not.toBeNull();
+  });
+
+  it('should add aria-label to support screen readers properly', async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+      <p-select-wrapper label="Some label">
+        <select name="some-name">
+          <option value="a">Option A</option>
+          <option value="b">Option B</option>
+          <option value="c">Option C</option>
+        </select>
+      </p-select-wrapper>
+    `);
+    const select = await page.find('p-select-wrapper select');
+    expect(select.getAttribute('aria-label')).toBe('Some label');
   });
 
   it('should not render label if label prop is not defined but should render if changed programmatically', async () => {
@@ -151,5 +168,51 @@ describe('select-wrapper', () => {
     }
 
     expect((await getFakeSelect())).not.toHaveClass('p-select-wrapper__fake-select--disabled');
+  });
+
+  describe('hover state', () => {
+    const getBoxShadow = async (page: E2EPage) => {
+      const fakeSelect = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
+      const styles = await fakeSelect.getComputedStyle();
+      return styles.boxShadow;
+    };
+
+    it('should change box-shadow color when fake select is hovered', async () => {
+      const page = await newE2EPage();
+      await page.setContent(`<p-select-wrapper label="Some label">
+        <select name="some-name">
+          <option value="a">Option A</option>
+          <option value="b">Option B</option>
+          <option value="c">Option C</option>
+        </select>
+      </p-select-wrapper>`);
+
+      const fakeSelect = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
+
+      const initialBoxShadow = await getBoxShadow(page);
+
+      await fakeSelect.hover();
+
+      expect(await getBoxShadow(page)).not.toBe(initialBoxShadow);
+    });
+
+    it('should change box-shadow color of fake select when label text is hovered', async () => {
+      const page = await newE2EPage();
+      await page.setContent(`<p-select-wrapper label="Some label">
+        <select name="some-name">
+          <option value="a">Option A</option>
+          <option value="b">Option B</option>
+          <option value="c">Option C</option>
+        </select>
+      </p-select-wrapper>`);
+
+      const labelText = await page.find('p-select-wrapper >>> .p-select-wrapper__label-text');
+
+      const initialBoxShadow = await getBoxShadow(page);
+
+      await labelText.hover();
+
+      expect(await getBoxShadow(page)).not.toBe(initialBoxShadow);
+    });
   });
 });