+++ packages/components-js/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Divider`
 
+### Fixed
+- Support label text of form elements for Screen readers
+
 ## [1.1.2] - 2020-03-17
 
 ### Changed